ultrasonic sensor debugging, seems to work, but need to convert now

+++ src/UltrasonicDriver.h
@@ -17,7 +17,8 @@ public:
 
     /// pollSensor - Sends a pulse from the sensor
     int8_t pollSensor();
-    int32_t readDistance();
+
+    float readDistance();
     volatile bool handleInterrupt();
 
 private:
